Drop redundant api which also wasn't working
The issue was that stateChanged() was emitted, and only after that m_remoteConnected being set to true.
This resulted in nymea-app to try and send the Hello call right on the connected and the proxy lib would
bail out saying that the remote is not connected yet.

I started to fix it, but realized that state already has StateRemoteConnected, so this is really
just a duplicate without any additional value.