Don't use format string function with variable input
New gcc refuses to compile this as it may load to crashes
if the input variable contains a format string literal, e.g. %s
and there are no ... args provided to fill those in.

Being at it, removed the convertString() method as QByteArray already
provides a constData() method, so it's not needed.